Xfinaltests sucess

+++ account-logic/src/main/java/se/liu/ida/tdp024/account/logic/api/facade/TransactionLogicFacade.java
@@ -5,9 +5,9 @@ import se.liu.ida.tdp024.account.data.api.entity.Transaction;
 import java.util.List;
 
 public interface TransactionLogicFacade {
-    public String debitAccount(int id, int amount);
+    public String debitAccount(long id, int amount);
 
-    public String creditAccount(int id, int amount);
+    public String creditAccount(long id, int amount);
 
     public List getTransactions(long id);
 }
